+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-08-03 19:02:54
-// Commit: test(auth): implement widget integration tests for auth
+// Updated at: 2026-08-03 21:16:19
+// Commit: feat(auth): implement shimmer loading skeleton in auth
 
 class authComponentState {
-  final String commitRef = '3318';
-  final String lastMessage = 'test(auth): implement widget integration tests for auth';
-  final DateTime timestamp = DateTime.parse('2026-08-03T19:02:54+05:00');
+  final String commitRef = '3325';
+  final String lastMessage = 'feat(auth): implement shimmer loading skeleton in auth';
+  final DateTime timestamp = DateTime.parse('2026-08-03T21:16:19+05:00');
 }